traffic2

Speed limits and highway safety.

data("traffic2")

Format

A data frame with 108 observations.

year

1981 to 1989

month

1 to 12

totacc

statewide total accidents

fatacc

statewide fatal accidents

injacc

statewide injury accidents

pdoacc

property damage only accidents

ntotacc

noninterstate total acc.

nfatacc

noninterstate fatal acc.

ninjacc

noninterstate injur acc.

npdoacc

noninterstate property acc.

rtotacc

total acc. on rural 65 mph roads

rfatacc

fatal acc. on rural 65 mph roads

rinjacc

injury acc. on rural 65 mph roads

rpdoacc

property acc. on rural 65 mph roads

ushigh

acc. on U.S. highways

cntyrds

acc. on county roads

strtes

acc. on state routes

unem

state unemployment rate

spdlaw

=1 after 65 mph in effect

beltlaw

=1 after seatbelt law

wkends

number of weekend days in month

Source

Jeffrey M. Wooldrige (2006): Introductory Econometrics: A Modern Approach, 3rd ed., Thomson South-Western.

Details

Data used in: P.S. McCarthy (1994), “Relaxed Speed Limits and Highway Safety: New Evidence from California,” Economics Letters 46, 173--179. Professor McCarthy kindly provided the data.

Notes

Many states have changed maximum speed limits and imposed seat belt laws over the past 25 years. Data similar to those in traffic2 should be fairly easy to obtain for a particular state. One should combine this information with changes in a state’s blood alcohol limit and the passage of per se and open container laws.